requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic==2.6.1
email-validator>=2.0.0
openai==1.12.0
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.23
langchain-text-splitters>=1.0.0
python-dateutil==2.8.2
httpx==0.26.0

